// backend/src/enums/dao-commands.enum.ts
export enum DaoCommandsEnum {
  addRowInTable = 'addRowInTable',
  deleteRowInTable = 'deleteRowInTable',
  getRowByPrimaryKey = 'getRowByPrimaryKey',
  getRowsFromTable = 'getRowsFromTable',
  getTableForeignKeys = 'getTableForeignKeys',
  getTablePrimaryColumns = 'getTablePrimaryColumns',
  getTableStructure = 'getTableStructure',
  getTablesFromDB = 'getTablesFromDB',
  testConnect = 'testConnect',
  updateRowInTable = 'updateRowInTable',
  validateSettings = 'validateSettings',
  getIdentityColumns = 'getIdentityColumns',
  getReferencedTableNamesAndColumns = 'getReferencedTableNamesAndColumns',
}